minizip: Sync with contrib from zlib 1.2.4, document Godot changes
Our minizip version was quite similar to that in zlib 1.2.4's
contrib/minizip, so used it as a base for syncing.

The Godot-specific changes used in core/io were kept for now and marked
with 'GODOT start' and 'GODOT end' comments. They are also shown in the
godot-zlib-1.2.4-minizip-seek.patch file.